pyproject.toml
[tool.poetry]
name = "project-python12"
version = "0.1.0"
description = ""
authors = ["baha_2002 <baktiarzaksylykov22@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0"
pillow = "^10.1.0"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
